refactor: useContext 사용하여 ToDoStore로 분리 및 파일 구조 정리
